🔧Move Fourier Encoder to layers
