+++ src/genpp/models/fm/fm_cnn.py
@@ -1,4 +1,3 @@
-import math
 from abc import ABC, abstractmethod
 from collections.abc import Callable, Sequence
 
@@ -9,33 +8,8 @@ from flow_matching.path import CondOTProbPath
 from flow_matching.solver import ODESolver
 from omegaconf import DictConfig
 
-from genpp.models.layers import CropND, PixelEmbedder
+from genpp.models.layers import CropND, FourierEncoder, PixelEmbedder
 from genpp.models.utils import BaseModule
-
-
-class FourierEncoder(nn.Module):
-    """
-    Based on https://github.com/lucidrains/denoising-diffusion-pytorch/blob/main/denoising_diffusion_pytorch/karras_unet.py#L183
-    """
-
-    def __init__(self, dim: int):
-        super().__init__()
-        assert dim % 2 == 0
-        self.half_dim = dim // 2
-        self.weights = nn.Parameter(torch.randn(1, self.half_dim))
-
-    def forward(self, t: torch.Tensor) -> torch.Tensor:
-        """
-        Args:
-            t (torch.Tensor): [bs, 1, 1, 1]
-        Returns:
-            torch.Tensor: [bs, dim]
-        """
-        t = rearrange(t, "... -> (...) 1")  # [bs, 1]
-        freqs = t * self.weights * 2 * math.pi  # [bs, half_dim]
-        sin_embed = torch.sin(freqs)  # [bs, half_dim]
-        cos_embed = torch.cos(freqs)  # [bs, half_dim]
-        return torch.cat([sin_embed, cos_embed], dim=-1) * math.sqrt(2)  # [bs, dim]
 
 
 class ResidualLayer(nn.Module):
